+++ libft/ft_memmove.c
@@ -6,15 +6,15 @@
 /*   By: ede-nada <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/31 19:29:28 by ede-nada          #+#    #+#             */
-/*   Updated: 2020/01/31 20:28:59 by ede-nada         ###   ########.fr       */
+/*   Updated: 2020/02/11 23:03:00 by evandrosu        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memmove(void *dst, const void *src, unsigned long int len)
+void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	unsigned long int i;
+	size_t i;
 
 	if (!dst && !src)
 		return (0);
